Use TPL DataFlow to better handle large batches (ordering and concurrency). Fix large sample count progress reporting.

+++ AudioWorks/src/AudioWorks.Api/ProgressToken.cs
@@ -19,6 +19,6 @@ namespace AudioWorks.Api
         /// Gets the total # of frames completed since the activity started.
         /// </summary>
         /// <value>The # of frames completed.</value>
-        public int FramesCompleted { get; internal set; }
+        public long FramesCompleted { get; internal set; }
     }
 }
